+++ src/main/enum/errors-message.enum.ts
@@ -1,0 +1,6 @@
+export enum ErrorsMessage {
+  requestError = "Error! Check if the URL is correct",
+  deploymentError = "Check if the Url is valid or if the camunda is running",
+  slashEndingUrl = "Url should not end with a slash /",
+  missingHttpProtocol = "Url should start with http:// or https://",
+}