+++ src/main/enum/literals.enum.ts
@@ -1,0 +1,5 @@
+export enum Literals {
+  newUrl = "$(ports-forward-icon) Add a new Camunda URL",
+  cleanUrlList = "$(notebook-delete-cell) Delete all Camunda Urls",
+  camundaUrlPlaceholder = "Camunda URL",
+}